chore: 未使用FilterSelectコンポーネント削除 + devポート3000→3005に変更
- filter-select.tsx → filter-utils.ts にリネーム（buildFilterUrlのみ残す）
- FilterSelect コンポーネントと "use client" / useRouter を削除
- Web devポートを3005に変更（Open WebUI とのポート3000競合を解消）
- .env.example / CLAUDE.md のポート記載を更新
