+++ .env.example
@@ -5,7 +5,7 @@ GCP_PROJECT_ID="hr-system-487809"
 FIRESTORE_EMULATOR_HOST="127.0.0.1:8080"
 
 # NextAuth.js
-NEXTAUTH_URL="http://localhost:3000"
+NEXTAUTH_URL="http://localhost:3005"
 NEXTAUTH_SECRET="generate-with-openssl-rand-base64-32"
 
 # Google OAuth
